Test facet results.

+++ common/src/main/java/io/sphere/sdk/search/TermFacetResult.java
@@ -5,12 +5,12 @@ import io.sphere.sdk.models.Base;
 import java.util.List;
 
 public class TermFacetResult extends Base implements FacetResult {
-    private final int missing;
-    private final int total;
-    private final int other;
+    private final long missing;
+    private final long total;
+    private final long other;
     private final List<TermStats> terms;
 
-    private TermFacetResult(final int missing, final int total, final int other, final List<TermStats> terms) {
+    private TermFacetResult(final long missing, final long total, final long other, final List<TermStats> terms) {
         this.missing = missing;
         this.total = total;
         this.other = other;
@@ -23,7 +23,7 @@ public class TermFacetResult extends Base implements FacetResult {
      * it would represent the amount of variants with no color associated.
      * @return amount of resources missing a value for the faceted field.
      */
-    public int getMissing() {
+    public long getMissing() {
         return missing;
     }
 
@@ -31,7 +31,7 @@ public class TermFacetResult extends Base implements FacetResult {
      * The number of resources matching some term in the facet.
      * @return amount of resources that match at least one term in the facet.
      */
-    public int getTotal() {
+    public long getTotal() {
         return total;
     }
 
@@ -39,7 +39,7 @@ public class TermFacetResult extends Base implements FacetResult {
      * The number of resources not included in the returned list of terms.
      * @return amount of resources matching the terms that could not be included in the facet result.
      */
-    public int getOther() {
+    public long getOther() {
         return other;
     }
 
@@ -51,7 +51,7 @@ public class TermFacetResult extends Base implements FacetResult {
         return terms;
     }
 
-    public static TermFacetResult of(final int missing, final int total, final int other, final List<TermStats> terms) {
+    public static TermFacetResult of(final long missing, final long total, final long other, final List<TermStats> terms) {
         return new TermFacetResult(missing, total, other, terms);
     }
 }
